requirements.txt
gTTS~=2.5.3
vosk==0.3.45
numpy==1.26.4
pillow~=10.4.0
word2number==1.1
sounddevice==0.4.6
customtkinter~=5.2.2